requirements.txt
ntplib
#pywintypes
python-dateutil